added packetgrabbing.lua
This is a simple packet grabber that defines time/protocol/source ip/destination ip/ source port/destination port. Line 9 requires you to add your path to where you want "captured_packets.txt". This is where the program will write the captured packets to. Also line 15 allows for max packet change, this is for testing so feel free to modify it. Also feel free to modify any of the a fields you want to capture at the top lines 1-6 but make sure you modify the packet_info function aswell as line 49 where it rights the information.