use pyshark to capture packets from Wi-Fi live
Takes live Wi-Fi information and saves a number of text files. 

Methods implemented:
live_packet_capture: uses pyshark to capture packets and returns a string with the packet information. It takes an interface, a given number of packets, and a filter. 
create_binary_files_and_directories: not currently used. Can be used to automatically save the wireshark packet capture as a wireshark file. takes a filepath
create_files_and_directories: used for text files, safely creates a text file. Takes a file path. 
savePacketInfo: Takes a file_path and a string of packet_info and writes it to the file. 
bigPacketCathcher: makes a number of text files with a given number of packets and a save path. It then saves the live packets in text files until it fills the given number of files with the given number of packets each. 

The formatting for the packet information is very versatile and is housed in the live_packet_capture